appback/2fa/requirements.txt
flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
PyJWT==2.7.0